fix help text and move help function to its own file
