+++ clang/tools/offload-arch/OffloadArch.h
@@ -13,4 +13,6 @@ std::string _aot_intelhd_capabilities(uint16_t vid, uint16_t devid,
                                      std::string oa);
 std::vector<std::string> _aot_get_requirements_from_file(std::string filename);
 
+void aot_usage();
+
 #endif
